Separated implementations for quantized & non-quantized tensors in reflection_pad2d_cpu
Summary: This PR is part of a series of PRs addressing https://github.com/pytorch/pytorch/issues/54150,
related to using dispatcher for calls to quantized backends as opposed to if/else conditionals.
This particular PR separates the calls to quantized & non-quantized backends for reflection_pad2d_cpu
using a dispatcher.

ghstack-source-id: 2f425b095882b72957fbc60328de476f21cb04ed
Pull Request resolved: https://github.com/pytorch/pytorch/pull/72442